requirements.txt
streamlit-authenticator
pickle
pathlib       